backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "google-generativeai>=0.8.5",
    "google-search-results>=2.4.2",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
    "google-api-python-client>=2.110.0",
    "reportlab>=3.6.12",
    "pytest>=8.2.0",
    "requests>=2.32.0",
    "httpx>=0.27.0",
]
